requirements.txt
boto3==1.40.62
pinecone==7.3.0
python-docx==1.2.0
pypdf==6.1.3
openai==2.6.1
python-dotenv==1.2.1
